Fix an issue where, previous Save/Don't save confirmation modal on Edit mode was showing in subsequent edits.
By setting the key to the full path of the file being previewed, we ensure that every time we open a different file, a completely new PreviewModal is created. This guarantees that all internal states, including the one causing the "Unsaved Changes" modal to reappear, are initialized correctly, eliminating issues with ghost modals reappearing.